chore(onboarding): split onboarding steps into own components
